Refactored and restructured project
Passing everything by params was getting messy and way to big so I
created two structs, Totp and Hotp and converted the related
functions to take those structs as a receiver.

Merged totp and hotp into a single otp package.

Added NewTOTP/HOTP func to create the struct with default values. I'm
not super happy with this method, I did some browsing and saw another
project that creates the structs and calls a setDefaults func at the
begining of each of it's receiver calls (hgfischer/go-otp). This person
seems to be an experienced Go dev and it might be a good idea to take
some notes from their design...